separate banners logic and single banner display
